AboveTustin - add tests to catch errors and assist with debugging

+++ bash/extras/abovetustin.sh
@@ -266,18 +266,32 @@ if [[ "${PHANTOMJS_EXISTS}" = "false" ]] ; then
         esac
 
         # Extract the files from the PhantomJS archive which was just downloaded.
+        if [[ -f "phantomjs-${PHANTOMJS_VERSION}-linux-${CPU_ARCHITECTURE}.tar" ]] ; then
         echo -e "\e[94m  Extracting the PhantomJS binary archive...\e[97m"
         echo ""
-        tar -vxfj phantomjs-${PHANTOMJS_VERSION}-linux-${CPU_ARCHITECTURE}.tar 2>&1
+        tar -vxj -f phantomjs-${PHANTOMJS_VERSION}-linux-${CPU_ARCHITECTURE}.tar 2>&1
         rm -vf phantomjs-${PHANTOMJS_VERSION}-linux-${CPU_ARCHITECTURE}.tar 2>&1
+        else
+            echo -e "\e[94m  Unable to extract the PhantomJS binary archive...\e[97m"
+        fi
 
         # Move the binary into the /usr/bin directory and make it executable.
+        if [[ -f "phantomjs-${PHANTOMJS_VERSION}-linux-${CPU_ARCHITECTURE}/bin/phantomjs" ]] ; then
         echo -e "\e[94m  Copying the PhantomJS binary into the directory /usr/bin...\e[97m"
         sudo cp -v phantomjs-${PHANTOMJS_VERSION}-linux-${CPU_ARCHITECTURE}/bin/phantomjs /usr/bin 2>&1
         echo ""
+        else
+            echo -e "\e[94m  Unable to copying the PhantomJS binary into the directory /usr/bin...\e[97m"
+        fi
+
+        # Make the binary in /usr/bin executable.
+        if [[ -f "/usr/bin/phantomjs" ]] ; then
         echo -e "\e[94m  Making the file /usr/bin/phantomjs executable...\e[97m"
         sudo chmod -v +x /usr/bin/phantomjs 2>&1
         echo ""
+        else
+            echo -e "\e[94m  Unable to make the file /usr/bin/phantomjs executable...\e[97m"
+        fi
 
     else
